[refact] 상품 관리 페이지 리팩토링
- 상품 추가, 수정 페이지 컴포넌트 합침(중복 삭제)
- 카테고리, 속성 초기화 되는 부분 useEffect cleanup함수로 변경
