+++ frontend/src/components/AdminProductAttr/index.tsx
@@ -1,15 +1,16 @@
 import { Fragment, useRef, useState } from 'react';
 import { Alert, CloseButton, Col, Form, Row, Table } from 'react-bootstrap';
 
-import { saveAttributeThunk } from '@redux/modules/categorySlice/thunk';
 import { useAppDispatch, useAppSelector } from '@hooks/reduxHooks';
+import { saveAttributeThunk } from '@redux/modules/categorySlice/thunk';
 import { setAttributesFromDb, setAttributesTable } from '@redux/modules/productSlice';
 
-const EditAttrs = () => {
+const AdminProductAttr = () => {
+  const dispatch = useAppDispatch();
   const selectedCategory = useAppSelector(state => state.category.selectedCategory);
   const attributesFromDb = useAppSelector(state => state.product.attributesFromDb);
   const attributesTable = useAppSelector(state => state.product.attributesTable);
-  const dispatch = useAppDispatch();
+
   const [newAttrKey, setNewAttrKey] = useState('');
   const [newAttrValue, setNewAttrValue] = useState('');
 
@@ -20,8 +21,8 @@ const EditAttrs = () => {
 
   const setValuesForAttrFromDbSelectForm = (e: React.ChangeEvent<HTMLSelectElement>) => {
     if (e.target.value !== '속성 선택') {
-      var selectedAttr = attributesFromDb?.find((item: any) => item.key === e.target.value);
-      let valuesForAttrKeys = attrVal.current;
+      const selectedAttr = attributesFromDb?.find((item: any) => item.key === e.target.value);
+      const valuesForAttrKeys = attrVal.current;
       if (selectedAttr && selectedAttr.value.length > 0) {
         while (valuesForAttrKeys.options.length) {
           valuesForAttrKeys.remove(0);
@@ -39,6 +40,11 @@ const EditAttrs = () => {
     if (e.target.value !== '속성 값 선택') {
       setAttributesTableWrapper(attrKey.current.value, e.target.value);
     }
+  };
+
+  const setAttributesTableWrapper = (key: string, val: string) => {
+    const modifiedTable = modifyAttributesTable(key, val);
+    dispatch(setAttributesTable(modifiedTable));
   };
 
   const modifyAttributesTable = (key: string, val: string) => {
@@ -59,11 +65,7 @@ const EditAttrs = () => {
     }
   };
 
-  const setAttributesTableWrapper = (key: string, val: string) => {
-    const newAttributesTable = modifyAttributesTable(key, val);
-    dispatch(setAttributesTable(newAttributesTable));
-  };
-
+  // Add new attribute
   const newAttrKeyHandler = (e: React.KeyboardEvent<HTMLInputElement>) => {
     e.preventDefault();
     setNewAttrKey(e.currentTarget.value);
@@ -101,9 +103,10 @@ const EditAttrs = () => {
     return attributesTable.filter((item: any) => item.key !== key);
   };
 
+  // Delete Attribute
   const deleteAttribute = (key: string) => {
-    const newAttributesTable = filterAttributesTable(key);
-    dispatch(setAttributesTable(newAttributesTable));
+    const modifiedTable = filterAttributesTable(key);
+    dispatch(setAttributesTable(modifiedTable));
   };
 
   return (
@@ -129,23 +132,24 @@ const EditAttrs = () => {
             </Form.Group>
           </Col>
           <Col md={6}>
-            <Form.Group className="mb-3" controlId="formBasicAttributeValue">
-              <Form.Label>속성 값</Form.Label>
+            <Form.Group className="mb-3" controlId="formBasicAttributes">
+              <Form.Label>속성 값 추가</Form.Label>
               <Form.Select
                 name="atrrVal"
                 aria-label="Default select example"
                 ref={attrVal}
                 onChange={attributeValueSelected}
               >
-                <option>속성 값 선택</option>
+                <option value="">속성 값 선택</option>
               </Form.Select>
             </Form.Group>
           </Col>
         </Row>
       )}
 
+      {/* 속성 테이블 */}
       <Row>
-        {attributesTable && attributesTable.length > 0 && (
+        {attributesTable?.length > 0 && (
           <Table hover>
             <thead>
               <tr>
@@ -155,7 +159,7 @@ const EditAttrs = () => {
               </tr>
             </thead>
             <tbody>
-              {attributesTable.map((item, idx) => (
+              {attributesTable?.map((item, idx) => (
                 <tr key={idx}>
                   <td>{item.key}</td>
                   <td>{item.value}</td>
@@ -180,7 +184,7 @@ const EditAttrs = () => {
               placeholder="먼저 카테고리를 선택하거나 생성하세요"
               ref={createNewAttrKey}
               disabled={['', 'Choose category'].includes(selectedCategory)}
-              required={!!newAttrValue}
+              required={!!newAttrKey}
               onKeyUp={newAttrKeyHandler}
             />
           </Form.Group>
@@ -194,13 +198,12 @@ const EditAttrs = () => {
               placeholder="먼저 카테고리를 선택하거나 생성하세요"
               ref={createNewAttrVal}
               disabled={['', 'Choose category'].includes(selectedCategory)}
-              required={!!newAttrKey}
+              required={!!newAttrValue}
               onKeyUp={newAttrValueHandler}
             />
           </Form.Group>
         </Col>
       </Row>
-
       <Alert show={!!newAttrKey && !!newAttrValue} variant="primary">
         속성 키와 속성 값이 모두 입력되어야 속성이 추가됩니다.
       </Alert>
@@ -208,4 +211,4 @@ const EditAttrs = () => {
   );
 };
 
-export default EditAttrs;
+export default AdminProductAttr;
