+++ frontend/src/pages/admin/Products/index.tsx
@@ -2,15 +2,11 @@ import { useState, useEffect } from 'react';
 import { Button, Table } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
 
-import { useAppDispatch } from '@hooks/reduxHooks';
-import { resetProductState } from '@redux/modules/productSlice';
-import { resetCategoryState } from '@redux/modules/categorySlice';
 import { getProductsForAdmin, deleteProduct } from '@utils/api';
 
 import AdminLayout from '@layout/AdminLayout';
 
 const ProductsPage = () => {
-  const dispatch = useAppDispatch();
   const [products, setProducts] = useState([]);
   const [productDeleted, setProductDeleted] = useState(false);
 
@@ -38,12 +34,6 @@ const ProductsPage = () => {
 
     return () => abortController.abort();
   }, [productDeleted]);
-
-  // reset attributes table
-  useEffect(() => {
-    dispatch(resetProductState());
-    dispatch(resetCategoryState());
-  }, []);
 
   return (
     <AdminLayout>
